Update README.md
Update readme to remove blurb about production usage. I think that using semver is enough of a warning to people that this is obviously not a 1.0, and  knowing what we know about semver, we will responsibly introduce changes. Let them decide if they want to use it for production. Case in point, React was heavily used in production before 1.0